Moves retries into InstrumentationHelper.
